feat: Ability to configure the `origin` of an entity
It is recommended to add information about the origin of MQTT entities by including the origin option (abbreviated as o) in the discovery payload. For device-based discovery, this information is required. The origin details will be logged in the core event log when an item is discovered or updated. Adding origin information helps with troubleshooting and provides valuable context about the source of MQTT messages in your Home Assistant setup.